Week 2 - Guessing Game
Learning how to properly make classes. This includes the differences
between classes, instances, and all the fun stuff that comes from that.
Basic OOP stuff, for better or worse.
Specifically, this week had us make a class which contained a few
attributes and methods that modified or reacted to those attributes. A
dummy class with a main was then used to create an instance of the first
class and make a game of sorts with that object.