Week 3 - TicTacToe
Intended as a demonstration of Java two-dimentional arrays, I also used
it to explore the power of enums. As this project required calling
functions within functions and passing status between them, using an
enum to ensure that all possible statuses are handled is quite nice.
I picked this coding pattern up in my exploration of the Rust
programming language - there, the compiler throws a full-on error if you
use an enum without handling all the possibilities. Eclipse only gives
you a warning, but that is still better than passing around nebulous
strings and using a billion if statements to figure out what you want.
Plus, using an enum is very reader-friendly.
The checkForWinner function is a bit painful to look at, but I'm not
sure how it could be done better. May warrant a rework in the future
(when I'm better at this programming stuff), but for now it works.