Week 6 - Discussion Example
This week's discussion post prompt wanted us to provide an example of a
superclass and at least two subclasses. It wasn't a requirement, but
this prompt just begged to have actual code for the example. So that's
what I did.
I've already done the discussion post, but I just now realized that I
could have enhanced the polymorphism part of it if I had made the cost
for grocery items be calculated using the weight... Oh well, something
to do better from the start the next time.