Week 7 - Stack Calculator
This week's topic was data structures, and thus the lab was working with
a stack. Nothing too fancy, just some basic math calculations using the
Stack java.util class.
I cannot wait until we're working with a GUI so that buttons can be used
to send enum-compatible commands without having to do big if/else if
parsing of strings commands.