Week 8 - Course Project Phase 3
This week involved working with superclasses and interfaces. Nothing
terribly difficult, but definitely time consuming to implement. For my
own benefit, I also refactored the school work directory a bit and
pulled some of the BankAcctApp Main out into a support method.

Specifically, I made this week's work a copy of Phase 2, as the changes
here were going to be fairly sweeping and I didn't want to risk deleting
something at the start I would need later and have to figure out how to
redo. I also moved our the user input validation method for creating a
new customer, as that was well over 100 lines of basically the same
thing over and over again with small variations for each of the required
data. There's probably a more efficient way to do that, but I don't know
what it is right now.

I am very much so looking forward to making a GUI if for no other reason
than to cut down on the superfluous info required to make getting user
input via the console reasonable. That being said, translating all of
these println's is going to be quite the process. Definitely going to
need to get an early start on that week's work...