Week 10 - Course Project Phase 4
The last assignment for CIS407, and also the one that probably took me
the longest. Building the GUI from scratch was quite the process, and I
was only able to pull a little bit of code from the previous iteration.
I am, however, quite proud of how it turned out. It's not perfect, by
any means, but it is more than sufficient for the class and still pretty
good by my standards.

I've also added a README.md, as this will likely be the last commit to
this repository. I waffle on enough it that, so there's not much more I
can add here. Except for how depressing/rewarding it was to delete the
UltimateStringTabber class and then implement an HTML table function
that does it so much better. I had been looking forward to making this
GUI for a lot of reasons, but discovering that Java supports displaying
HTML with (some) CSS definitely added to that list.